# api/openapi-v1.yaml
openapi: 3.0.2
info:
  title: Facade
  description: This is the public API for Facade.
  version: 0.0.1
servers:
  - url: https://facade.carpedeez.io/v0
paths:
  /@me:
    get:
      summary: Get Session
      operationId: me
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        default:
          description: Unexpected Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /assets:
    post:
      summary: Upload file
      operationId: uploadFile
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Upload'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: string
                description: URL for image
        default:
          description: Unexpected Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /u:
    post:
      summary: Create user
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostUser'
      responses:
        '204':
          description: OK
        default:
          description: Unexpected Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /u/{username}:
    get:
      summary: Get user
      operationId: getUser
      parameters:
        - name: username
          required: true
          in: path
          schema:
            type: string
            minLength: 3
            maxLength: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetUser'
        default:
          description: Unexpected Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      summary: Update user
      operationId: updateUser
      parameters:
        - name: username
          required: true
          in: path
          schema:
            type: string
            minLength: 3
            maxLength: 20
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchUser'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetUser'
        default:
          description: Unexpected Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /d:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostDisplay'
      summary: Create display
      operationId: createDisplay
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: integer
                format: int64
        '400':
          description: Bad Request
        default:
          description: Unexpected Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /d/{displayID}:
    get:
      summary: Get display
      operationId: getDisplay
      parameters:
        - name: displayID
          required: true
          in: path
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetDisplay'
        default:
          description: Unexpected Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      summary: Update display
      operationId: updateDisplay
      parameters:
        - name: displayID
          required: true
          in: path
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchDisplay'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetDisplay'
        default:
          description: Unexpected Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete display
      operationId: deleteDisplay
      parameters:
        - name: displayID
          required: true
          in: path
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: No Content
        default:
          description: Unexpected Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /i:
    post:
      summary: Create Item
      operationId: createItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostItem'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetItem'
        default:
          description: Unexpected Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /i/{itemID}:
    get:
      summary: Get item
      operationId: getItem
      parameters:
        - name: itemID
          required: true
          in: path
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetItem'
        default:
          description: Unexpected Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      summary: Update item
      operationId: updateItem
      parameters:
        - name: itemID
          required: true
          in: path
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchItem'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetItem'
        default:
          description: Unexpected Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete item
      operationId: deleteItem
      parameters:
        - name: itemID
          required: true
          in: path
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
        default:
          description: Unexpected Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    Session:
      type: object
      required:
        - id
        - active
        - expires_at
        - authenticated_at
        - authenticator_assurance_level
        - authentication_methods
        - issued_at
        - identity
      properties:
        id:
          type: string
        active:
          type: boolean
        expires_at:
          type: string
        authenticated_at:
          type: string
        authenticator_assurance_level:
          type: string
        authentication_methods:
          type: array
          items:
            type: object
            required:
              - method
              - completed_at
            properties:
              method:
                type: string
              completed_at:
                type: string
        issued_at:
          type: string
        identity:
          type: object
          required:
            - id
            - schema_id
            - schema_url
            - state
            - state_changed_at
            - traits
            - verifiable_addresses
            - recovery_addresses
            - created_at
            - updated_at
          properties:
            id:
              type: string
            schema_id:
              type: string
            schema_url:
              type: string
            state:
              type: string
            state_changed_at:
              type: string
            traits:
              type: object
              required:
                - website
                - email
              properties:
                website:
                  type: string
                email:
                  type: string
            verifiable_addresses:
              type: array
              items:
                type: object
                required:
                  - id
                  - value
                  - verified
                  - via
                  - status
                  - created_at
                  - updated_at
                properties:
                  id:
                    type: string
                  value:
                    type: string
                  verified:
                    type: boolean
                  via:
                    type: string
                  status:
                    type: string
                  created_at:
                    type: string
                  updated_at:
                    type: string
            recovery_addresses:
              type: array
              items:
                type: object
                required:
                  - id
                  - value
                  - via
                  - created_at
                  - updated_at
                properties:
                  id:
                    type: string
                  value:
                    type: string
                  via:
                    type: string
                  created_at:
                    type: string
                  updated_at:
                    type: string
            created_at:
              type: string
            updated_at:
              type: string
    Upload:
      required:
        - component
        - file
      type: object
      additionalProperties: false
      properties:
        component:
          type: string
        file:
          type: string
          format: binary
    GetUser:
      type: object
      required:
        - id
        - username
        - firstName
        - lastName
        - displayIDs
        - photoURL
        - socialLinks
      additionalProperties: false
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        displayIDs:
          type: array
          items:
            type: integer
            format: int64
        photoURL:
          type: string
        socialLinks:
          type: array
          items:
            type: string
    PostUser:
      description: registration
      type: object
      required:
        - username
        - firstName
        - lastName
      additionalProperties: false
      properties:
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
    PatchUser:
      type: object
      additionalProperties: false
      properties:
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        socialLinks:
          type: array
          items:
            type: string
    GetDisplay:
      type: object
      required:
        - id
        - username
        - title
        - description
        - photoURL
        - itemIDs
      additionalProperties: false
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        title:
          type: string
        description:
          type: string
        photoURL:
          type: string
        itemIDs:
          type: array
          items:
            type: integer
            format: int64
    PostDisplay:
      type: object
      required:
        - title
        - description
      additionalProperties: false
      properties:
        title:
          type: string
        description:
          type: string
    PatchDisplay:
      type: object
      additionalProperties: false
      properties:
        title:
          type: string
        description:
          type: string
    GetItem:
      type: object
      required:
        - id
        - externalLink
        - socialPostLink
        - photoURL
        - username
        - displayID
      additionalProperties: false
      properties:
        id:
          type: integer
          format: int64
        externalLink:
          type: string
        socialPostLink:
          type: string
        photoURL:
          type: string
        username:
          type: string
        displayID:
          type: integer
          format: int64
    PostItem:
      type: object
      required:
        - externalLink
        - socialPostLink
      additionalProperties: false
      properties:
        externalLink:
          type: string
        socialPostLink:
          type: string
    PatchItem:
      type: object
      additionalProperties: false
      properties:
        externalLink:
          type: string
        socialPostLink:
          type: string
        photoURL:
          type: string
    Error:
      type: object
      required:
        - code
        - message
      additionalProperties: false
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string